Moved TeamHome pages to subfolder
